Fix warning
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@10496 3db68df2-c116-0410-a063-a993310a9797

+++ source/ps/Profiler2.cpp
@@ -305,7 +305,7 @@ void CProfiler2::ThreadStorage::RecordAttribute(const char* fmt, va_list argp)
 	// (Don't use vsprintf_s because it treats overflow as fatal)
 
 	// Terminate the string if the printing was truncated
-	if (len < 0 || len >= MAX_ATTRIBUTE_LENGTH - 1)
+	if (len < 0 || len >= (int)MAX_ATTRIBUTE_LENGTH - 1)
 	{
 		strncpy(buffer + 4 + MAX_ATTRIBUTE_LENGTH - 4, "...", 4);
 		len = MAX_ATTRIBUTE_LENGTH - 1; // excluding null terminator
